+++ lib/tests/integration/wed-decoration-test.ts
@@ -9,7 +9,6 @@ import * as sinon from "sinon";
 
 import { CaretManager } from "wed/caret-manager";
 import { isElement } from "wed/domtypeguards";
-import * as log from "wed/log";
 import * as onerror from "wed/onerror";
 import * as wed from "wed/wed";
 
@@ -90,7 +89,6 @@ describe("wed decoration:", () => {
 
     // We don't reload our page so we need to do this.
     onerror.__test.reset();
-    log.clearAppenders();
     expect(wasTerminating)
       .to.equal(false, "test caused an unhandled exception to occur");