+++ lib/tests/integration/wed-error-handling-test.ts
@@ -8,7 +8,6 @@ import * as mergeOptions from "merge-options";
 import * as sinon from "sinon";
 
 import * as keyConstants from "wed/key-constants";
-import * as log from "wed/log";
 import * as onerror from "wed/onerror";
 import * as wed from "wed/wed";
 
@@ -67,7 +66,6 @@ describe("wed error handling:", () => {
 
     // We don't reload our page so we need to do this.
     onerror.__test.reset();
-    log.clearAppenders();
     expect(wasTerminating)
       .to.equal(false, "test caused an unhandled exception to occur");