+++ lib/tests/integration/wed-file-state-test.ts
@@ -9,7 +9,6 @@ import * as sinon from "sinon";
 
 import { CaretManager } from "wed/caret-manager";
 import * as keyConstants from "wed/key-constants";
-import * as log from "wed/log";
 import * as onerror from "wed/onerror";
 import * as wed from "wed/wed";
 
@@ -79,7 +78,6 @@ describe("wed file state:", () => {
 
     // We don't reload our page so we need to do this.
     onerror.__test.reset();
-    log.clearAppenders();
     expect(wasTerminating)
       .to.equal(false, "test caused an unhandled exception to occur");