+++ lib/tests/integration/wed-paste-copy-cut-test.ts
@@ -9,7 +9,6 @@ import * as sinon from "sinon";
 
 import * as browsers from "wed/browsers";
 import { CaretManager } from "wed/caret-manager";
-import * as log from "wed/log";
 import * as onerror from "wed/onerror";
 import * as wed from "wed/wed";
 
@@ -91,7 +90,6 @@ describe("wed paste copy cut:", () => {
 
     // We don't reload our page so we need to do this.
     onerror.__test.reset();
-    log.clearAppenders();
     expect(wasTerminating)
       .to.equal(false, "test caused an unhandled exception to occur");