+++ lib/tests/integration/wed-typing-test.ts
@@ -12,7 +12,6 @@ import { isElement } from "wed/domtypeguards";
 import { childByClass, closestByClass, indexOf } from "wed/domutil";
 import * as key from "wed/key";
 import * as keyConstants from "wed/key-constants";
-import * as log from "wed/log";
 import * as onerror from "wed/onerror";
 import * as wed from "wed/wed";
 
@@ -93,7 +92,6 @@ describe("wed typing:", () => {
 
     // We don't reload our page so we need to do this.
     onerror.__test.reset();
-    log.clearAppenders();
     expect(wasTerminating)
       .to.equal(false, "test caused an unhandled exception to occur");