+++ lib/tests/integration/wed-without-saver-test.ts
@@ -6,7 +6,6 @@
 import { expect } from "chai";
 import * as sinon from "sinon";
 
-import * as log from "wed/log";
 import * as onerror from "wed/onerror";
 import * as wed from "wed/wed";
 
@@ -79,7 +78,6 @@ describe("wed without saver:", () => {
 
     // We don't reload our page so we need to do this.
     onerror.__test.reset();
-    log.clearAppenders();
     expect(wasTerminating)
       .to.equal(false, "test caused an unhandled exception to occur");